+++ ichnaea/data/tests/test_export.py
@@ -127,7 +127,7 @@ class TestExporter(BaseExportTest):
         self.celery_app.export_queues = queues = configure_export(
             self.redis_client, config)
         self.test_queue_key = queues['test'].queue_key()
-        self.session.add(ApiKey(valid_key='test2', log=True))
+        self.session.add(ApiKey(valid_key='test2', log_submit=True))
         self.session.flush()
 
     def test_enqueue_reports(self):
@@ -184,7 +184,7 @@ class TestGeosubmitUploader(BaseExportTest):
             self.redis_client, config)
 
     def test_upload(self):
-        self.session.add(ApiKey(valid_key='e5444-794', log=True))
+        self.session.add(ApiKey(valid_key='e5444-794', log_submit=True))
         self.session.flush()
 
         reports = []
@@ -245,7 +245,7 @@ class TestS3Uploader(BaseExportTest):
         self.assertFalse(export_queue.monitor_name)
 
     def test_upload(self):
-        self.session.add(ApiKey(valid_key='e5444-794', log=True))
+        self.session.add(ApiKey(valid_key='e5444-794', log_submit=True))
         self.session.flush()
 
         reports = self.add_reports(3, email='secretemail@localhost',